+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/graph/BindingGraphGenerator.kt
@@ -145,7 +145,7 @@ internal class BindingGraphGenerator(
           )
 
         graph.addInjector(contextKey, entry)
-        if (contextKey.typeKey in graph) {
+        if (contextKey.typeKey in bindingLookup) {
           // Injectors may be requested multiple times, don't double-register
           continue
         }
@@ -182,7 +182,9 @@ internal class BindingGraphGenerator(
         }
 
         val isInherited = typeKey in inheritedProviderFactoryKeys
-        if (!providerFactory.annotations.isIntoMultibinding && typeKey in graph && isInherited) {
+        if (
+          !providerFactory.annotations.isIntoMultibinding && typeKey in bindingLookup && isInherited
+        ) {
           // If we already have a binding provisioned in this scenario, ignore the parent's version
           continue
         }
@@ -254,7 +256,7 @@ internal class BindingGraphGenerator(
         val isInherited = typeKey in inheritedBindsCallableKeys
         if (
           !bindsCallable.callableMetadata.annotations.isIntoMultibinding &&
-            typeKey in graph &&
+            typeKey in bindingLookup &&
             isInherited
         ) {
           // If we already have a binding provisioned in this scenario, ignore the parent's version